Fix for #4: [FR] Support move sync, bumped version to 0.1.2
The edit history file was failing to be moved alongside the note when a note was
moved to a different folder.

This was due to code put in place to prevent history file renames/moves because,
when a folder is moved, Obsidian will move all the contents automatically, and
if the edit history file is moved explicitly before Obsidian moves it, then
Obsidian will throw a benign error when it tries to move the edit history file
and it doesn't find it.

The code was failing to allow the rename when a single note, but not the parent
folders, are moved across folders.

In summary the plugin should rename/move the edit history file when the note is
renamed or the direct parent folder changes, and let Obsidian move the edit
history file in all the other cases.

This fixes #4: [FR] Support move sync

Files:

README.md
- Minor clarification on history file format.

main.ts

- When there's a rename of a note, also move/rename the history file if either:
  - The note is changing name
  - The note is changing parent (but not any other ancestor)
- Added comments around
- Simplified link creation (no functional changes)

manifest.json
package.json
versions.json
- bumped to version 0.1.2

releases/0.1.2/main.js
releases/0.1.2/manifest.json
releases/0.1.2/styles.css
  - This release

Tests:
- Renamed note, verified edit history file is renamed by plugin
- Moved note to different folder, verified edit history file is moved by plugin
- Moved note to root folder, verified edit history file is moved by plugin
- Moved note from root folder, verified edit history file is moved by plugin
- Moved parent folder to different folder/root, verified edit history file is
  moved by Obsidian
- Moved grandparent to different folder/root, verified edit history file is
  moved by Obsidian